Update simple3.php
These are 115 lines off code that work together also works with mootools. Tested it on 800.000 lines of code and comments. works fine. 16-01-2016..! This is the code without the comments in it.!!!!    It works with REGEX and HTML code in javascript too. And with brackets and parenthetical and collons also like this (( testing( 'ajfhh "test2"'(This))))...and mutch more www.heijer-onderhoud.nl you can see code in action. It works on Joomla sites. So JQuery is working to and many more famous scripts. also with other cms website software like Drupal...